Better command and options handling
Deciding about fail or success of the command parsing based on the
returned value of the command parsers was highly impractical. Parsers
had to return the empty list on success and their arguments on failure
which then had to be evaluated by parse-command. Using continuations
made it possible to shift the control flow from the central parsing
manager to the specific parsing functions, eliminating a lot of argument
passing and boilerplate code.

The biggest problem with option handling was that the programs flow
could not be controlled. Specific options could not end the parsing
process and since command parsing was based on the command part of the
arguments, command parsing could not be prevented. By using 2
continuations, one being the command parser, one to continue the option
parsing process, single option parsers are enabled to terminate the
program without the need for an exit expression.