Delete multiple tag-lists & read stdin
Multiple tag-lists can now be passed on the command line or on standard
input. This also fixes the inconsistent behaviour of delete-tags in
comparison to tag and untag as it doesn't crash anymore on empty input.

To be able to spot errors easily, parse-tag-list was changed to emit
incorrect tag-lists.