Fix reversed tokens in parentheses
Fix a bug introduced in:
45fe8f22bece tokenize: cons-reverse instead of append

The tokenizer procedure was transformed to cons new tokens to the front of the
token-list and reverse it when returning. Because the recursive application of
the tokenizer when processing filters contained in parentheses did not reverse
the list when returning, the parts of filters contained in parentheses would be
reversed.

Simply reversing the list in the base case processing a closing parenthesis
resolves the issue.